Allow disabling and switching between heating/cooling live.

+++ Marlin/src/libs/autoreport.h
@@ -26,13 +26,13 @@
 template <typename Helper>
 struct AutoReporter {
   millis_t next_report_ms;
-  uint8_t report_interval;
+  float report_interval;
   #if HAS_MULTI_SERIAL
     SerialMask report_port_mask;
     AutoReporter() : report_port_mask(SerialMask::All) {}
   #endif
 
-  inline void set_interval(uint8_t seconds, const uint8_t limit=60) {
+  inline void set_interval(float seconds, const uint8_t limit=60) {
     report_interval = _MIN(seconds, limit);
     next_report_ms = millis() + SEC_TO_MS(seconds);
   }
